Move HexPainter to separate file
